+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## >= 8.0.0 (WIP)
 
-- Removed support for depecrated `extraAriaContext` (please use `ariaLabelBuilder` instead)
+- Providing the `forcePage` prop (to something else than `undefined`) now makes the component wholly [controlled](https://reactjs.org/docs/forms.html#controlled-components). (see https://github.com/AdeleD/react-paginate/issues/124)
 - Add a warning when providing both `initialPage` and `forcePage` props. You should only provides `forcePage` when the component is [controlled](https://reactjs.org/docs/forms.html#controlled-components). (see: https://github.com/AdeleD/react-paginate/pull/290)
+- Removed support for depecrated `extraAriaContext` (please use `ariaLabelBuilder` instead)
 
 ## >= 7.1.4